+++ example/ios/example/AppDelegate.swift
@@ -9,27 +9,32 @@ import UIKit
 import React
 import React_RCTAppDelegate
 import ReactAppDependencyProvider
+import SpotImCore
  
 @main
 class AppDelegate: RCTAppDelegate {
-  var window: UIWindow?
-  var navController: UINavigationController?
-
   override func application(_ application: UIApplication, didFinishLaunchingWithOptions launchOptions: [UIApplication.LaunchOptionsKey : Any]? = nil) -> Bool {
+    self.moduleName = "example"
+    self.dependencyProvider = RCTAppDependencyProvider()
+ 
+    // You can add your custom initial props in the dictionary below.
+    // They will be passed down to the ViewController used by React Native.
+    self.initialProps = [:]
+    
     let bridge = RCTBridge(delegate: self, launchOptions: launchOptions)
-    let rootView = RCTRootView(bridge: bridge!, moduleName: "example", initialProperties: nil)
+    let rootView = RCTRootView(bridge: bridge!, moduleName: self.moduleName!, initialProperties: nil)
 
-    let window = UIWindow(frame: UIScreen.main.bounds)
     let rootViewController = UIViewController()
     rootViewController.view = rootView
 
     let navController = UINavigationController(rootViewController: rootViewController)
     navController.setNavigationBarHidden(true, animated: false)
-    window.rootViewController = navController
-    window.makeKeyAndVisible()
 
-    self.window = window
-    self.navController = navController
+    self.window = UIWindow(frame: UIScreen.main.bounds)
+    self.window.rootViewController = navController
+    self.window.makeKeyAndVisible()
+    
+    GoogleAdsProvider.setSpotImSDKWithProvider()
 
     return true
   }
